Move TreePlacement into its own unit as well
